requirements.txt
numpy
pandas
scikit-learn
matplotlib
seaborn
einops
numba
scipy
icecream
xlrd
tomli-w
zero
category_encoders
imbalanced-learn
kaggle
transformers
datasets
peft
ml_collections
sdmetrics
prdc
rdt
openpyxl
xgboost
dgl
pydantic
pytorch_geometric
tqdm
